refactor: move command to module
