+++ .dockerignore
@@ -1,2 +1,10 @@
-.venv
-tests
+.venv/
+.git/
+tests/
+helm/
+.idea/
+.github/
+.vim/
+config/settings.local.json
+.mypy_cache/
+__pycache__/
