+++ tg_qso_bot/bot/handlers/__init__.py
@@ -1,0 +1,2 @@
+from .request_qso import request_qso
+
